Add files via upload
Data preparation for unsupervised learning with Python
We use the Iris dataset to easily understand the subject and estimate the data. There are 150 records in this data set that have 4 characteristics. These characteristics include petal length, petal width, sepal length, sepal width. Iris data set has 3 classes Setosa, Virginica and Versicolor. In this example, we will introduce all 4 characteristics of our flower to the unsupervised learning algorithm, and this model will recognize which class each iris belongs to.

In this example, the python scikit-learn library is used to load the Iris data and matplotlib is used to visualize the data.